+++ lib/Targets/AIETargetShared.cpp
@@ -89,7 +89,8 @@ void generateXAieDmaSetMultiDimAddr(raw_ostream &output, int ndims,
                                     int offsetA, int lenA,
                                     int elementWidthInBytes,
                                     const char *errorRetval) {
-  uint16_t elementWidthIn32bWords = elementWidthInBytes / 4;
+  double elementWidthIn32bWords =
+      static_cast<double>(elementWidthInBytes) / 4.0;
   std::string tensor = tileDMATensorStr(col, row, bdNum);
   output << "XAie_DmaTensor " << tensor << " = {};\n";
   output << tensor << ".NumDim = " << std::to_string(ndims) << ";\n";
@@ -108,7 +109,8 @@ void generateXAieDmaSetMultiDimAddr(raw_ostream &output, int ndims,
     // Assume AIE-ML architecture; we assert this above
     output << tensor << ".Dim[" << std::to_string(j) << "].AieMlDimDesc"
            << " = { /* StepSize */ "
-           << std::to_string(dims[i].getStride() * elementWidthIn32bWords)
+           << std::to_string(static_cast<uint32_t>(dims[i].getStride() *
+                                                   elementWidthIn32bWords))
            << ", /* Size */ " << std::to_string(dims[i].getSize()) << "};\n";
   }
   if ((baseAddrA + offsetA) % 4)
